alx_travel_app/requirements.txt
Django>=5.0
djangorestframework
psycopg2-binary
drf-yasg